feat(global): change design pattern for the whole project

+++ TMDB/app-modules/detail-module/entities/Review.swift
@@ -2,7 +2,7 @@
 //  Review.swift
 //  TMDB
 //
-//  Created by Mikhael Adiputra on 12/01/23.
+//  Created by Mikhael Adiputra on 02/02/23.
 //
 
 import Foundation
@@ -51,7 +51,7 @@ extension Review  {
 // MARK: - AuthorDetails
 struct AuthorDetails: Decodable {
     let name, username: String
-    let rating: Double?
+    let rating: Int?
 }
 
 extension AuthorDetails {
@@ -64,6 +64,6 @@ extension AuthorDetails {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         name    = try container.decode(String.self, forKey: .name)
         username   =  try container.decode(String.self, forKey: .username)
-        rating   =  try container.decode(Double?.self, forKey: .rating)
+        rating   =  try container.decode(Int?.self, forKey: .rating)
     }
 }
